+++ R/curve_line.R
@@ -40,7 +40,7 @@ curve_line <- S7::new_class(
   },
   constructor = function(x, y, trans = trans_identity(), ...) {
     S7::new_object(
-      drawable(geometry = "path", trans = trans),
+      drawable(geometry = "path", trans = trans, pathlike = TRUE),
       x = x,
       y = y,
       style = style(...)